Test de création de terrain

+++ src/Constants/Display.java
@@ -4,10 +4,10 @@ package Constants;
  * Constants for the display
  */
 public final class Display {
-    public static final int WINDOW_WIDTH = 1280;
-    public static final int WINDOW_HEIGHT = 720;
-    public static final int WINDOWED_WINDOW_WIDTH = 1024;
-    public static final int WINDOWED_WINDOW_HEIGHT = 576;
+    public static final double WINDOW_WIDTH = 1280;
+    public static final double WINDOW_HEIGHT = 720;
+    public static final double WINDOWED_WINDOW_WIDTH = 1024;
+    public static final double WINDOWED_WINDOW_HEIGHT = 576;
     public static final String GAME_NAME = "Atlantis (nom provisoire)";
 
     public static final double MAIN_BUTTON_X_SPACING = 200;
